rag: reuse `chunk_approximately_equals` util as prep for openai integration
