+++ sdks/python/apache_beam/ml/rag/embeddings/huggingface_test.py
@@ -27,6 +27,7 @@ from apache_beam.ml.rag.embeddings.huggingface import HuggingfaceTextEmbeddings
 from apache_beam.ml.rag.types import Chunk
 from apache_beam.ml.rag.types import Content
 from apache_beam.ml.rag.types import Embedding
+from apache_beam.ml.rag.test_utils import TestHelpers
 from apache_beam.ml.transforms.base import MLTransform
 from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.util import assert_that
@@ -38,19 +39,6 @@ try:
   SENTENCE_TRANSFORMERS_AVAILABLE = True
 except ImportError:
   SENTENCE_TRANSFORMERS_AVAILABLE = False
-
-
-def chunk_approximately_equals(expected, actual):
-  """Compare embeddings allowing for numerical differences."""
-  if not isinstance(expected, Chunk) or not isinstance(actual, Chunk):
-    return False
-
-  return (
-      expected.id == actual.id and expected.metadata == actual.metadata and
-      expected.content == actual.content and
-      len(expected.embedding.dense_embedding) == len(
-          actual.embedding.dense_embedding) and
-      all(isinstance(x, float) for x in actual.embedding.dense_embedding))
 
 
 @pytest.mark.uses_transformers
@@ -105,7 +93,8 @@ class HuggingfaceTextEmbeddingsTest(unittest.TestCase):
           with_transform(embedder))
 
       assert_that(
-          embeddings, equal_to(expected, equals_fn=chunk_approximately_equals))
+          embeddings,
+          equal_to(expected, equals_fn=TestHelpers.chunk_approximately_equals))
 
 
 if __name__ == '__main__':
